requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-huggingface==0.3.1
streamlit
python-dotenv
chromadb==1.0.20
sentence-transformers==5.1.0
langchain_chroma-0.2.5